+++ src/main/java/blackjack/domain/card/Symbol.java
@@ -1,0 +1,39 @@
+package blackjack.domain.card;
+
+public enum Symbol {
+    ACE("A", 1),
+    TWO("2", 2),
+    THREE("3", 3),
+    FOUR("4", 4),
+    FIVE("5", 5),
+    SIX("6", 6),
+    SEVEN("7", 7),
+    EIGHT("8", 8),
+    NINE("9", 9),
+    TEN("10", 10),
+    JACK("J", 10),
+    QUEEN("Q", 10),
+    KING("K", 10);
+
+    private static final int ACE_LARGE_VALUE = 11;
+
+    private final String name;
+    private final int value;
+
+    Symbol(String name, int value) {
+        this.name = name;
+        this.value = value;
+    }
+
+    public static int getAceWeight() {
+        return ACE_LARGE_VALUE - ACE.value;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getValue() {
+        return value;
+    }
+}
