+++ src/main/java/blackjack/domain/card/Type.java
@@ -1,0 +1,18 @@
+package blackjack.domain.card;
+
+public enum Type {
+    DIAMOND("다이아몬드"),
+    HEART("하트"),
+    SPADE("스페이드"),
+    CLUB("클로바");
+
+    private final String name;
+
+    Type(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
