--- src/main/java/domain/GameResult.java
@@ -1,52 +1,0 @@
-package domain;
-
-import domain.user.Dealer;
-import domain.user.Player;
-import domain.user.Players;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-public class GameResult {
-
-    private final Map<Player, Outcome> playerResults;
-    private final Map<Outcome, Integer> dealerResults;
-
-    public GameResult(Dealer dealer, Players players) {
-        this.playerResults = Collections.unmodifiableMap(calculatePlayerResults(dealer, players));
-        this.dealerResults = Collections.unmodifiableMap(calculateDealerResults());
-    }
-
-    private Map<Player, Outcome> calculatePlayerResults(Dealer dealer, Players players) {
-        Map<Player, Outcome> playerResults = new LinkedHashMap<>();
-        for (Player player : players.getPlayers()) {
-            playerResults.put(player, player.calculateOutcome(dealer.getScore()));
-        }
-        return playerResults;
-    }
-
-    private Map<Outcome, Integer> calculateDealerResults() {
-        Map<Outcome, Integer> dealerResults = initializationDealerResults();
-        for (Outcome outcome : this.playerResults.values()) {
-            Outcome dealerOutcome = Outcome.converseOutcome(outcome);
-            dealerResults.put(dealerOutcome, dealerResults.get(dealerOutcome) + 1);
-        }
-        return dealerResults;
-    }
-
-    private Map<Outcome, Integer> initializationDealerResults() {
-        Map<Outcome, Integer> dealerResults = new LinkedHashMap<>();
-        for(Outcome outcome : Outcome.values()) {
-            dealerResults.put(outcome, 0);
-        }
-        return dealerResults;
-    }
-
-    public Map<Player, Outcome> getPlayersResult() {
-        return playerResults;
-    }
-
-    public Map<Outcome, Integer> getDealerResults() {
-        return dealerResults;
-    }
-}
