--- src/main/java/domain/user/Player.java
@@ -1,26 +1,0 @@
-package domain.user;
-
-import domain.Outcome;
-import util.BlackJackRule;
-
-public class Player extends User {
-
-    public Player(String name) {
-        super(name);
-    }
-
-    @Override
-    public boolean canDrawCard() {
-        return !(cards.isBust() || cards.isBlackJack());
-    }
-
-    public Outcome calculateOutcome(int dealerScore) {
-        if (cards.isBust()) {
-            return Outcome.LOSE;
-        }
-        if (BlackJackRule.isBust(dealerScore)) {
-            return Outcome.WIN;
-        }
-        return Outcome.calculate(dealerScore, cards.getScore());
-    }
-}
