--- src/main/java/domain/user/Players.java
@@ -1,30 +1,0 @@
-package domain.user;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class Players {
-
-    private final List<Player> players;
-
-    public Players(List<String> playerNames) {
-        List<Player> players = new ArrayList<>();
-        for (String playerName : playerNames) {
-            players.add(new Player(playerName));
-        }
-        this.players = Collections.unmodifiableList(players);
-    }
-
-    public List<Player> getPlayers() {
-        return players;
-    }
-
-    public List<String> getPlayerNames() {
-        List<String> playerNames = new ArrayList<>();
-        for (Player player : players) {
-            playerNames.add(player.getName());
-        }
-        return Collections.unmodifiableList(playerNames);
-    }
-}
